+++ packages/ui/src/components/dashboard/aeria-toast/aeria-toast.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import type { PhosphorIcon } from 'waltz-icons/core'
+import type { PhosphorIcon } from '@phosphor-icons/core'
 import { useStore } from '@waltz-ui/state-management'
 import AeriaIcon from '../../aeria-icon/aeria-icon.vue'