+++ packages/i18n/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @waltz-ui/i18n
 
+## 1.0.7
+
+### Patch Changes
+
+- Enforce typed icons
+
 ## 1.0.6
 
 ### Patch Changes